move `type_check` out of `compute_regions`
